backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pydantic>=2.5.0,<3.0.0
google-generativeai
sentence-transformers
faiss-cpu
streamlit
python-dotenv
Pillow
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-socketio==5.4.0
python-dotenv==0.19.0
websockets==10.0 
torch
transformers
accelerate
pillow
pix2tex 
onnxruntime-gpu
